+++ core/src/main/java/com/szpzs/model/Picture.java
@@ -2,7 +2,7 @@ package com.szpzs.model;
 
 import java.io.Serializable;
 import javax.persistence.*;
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 
 /**
@@ -13,25 +13,25 @@ import java.math.BigDecimal;
 @Table(name="PICTURES")
 @NamedQuery(name="Picture.findAll", query="SELECT p FROM Picture p")
 public class Picture implements Serializable {
-	private static final long serialVersionUID = 1L;
+	private static final Long serialVersionUID = 1L;
 
 	@Id
-	private long id;
+	private Long id;
 
 	@Column(name="\"LINK\"")
 	private String link;
 
 	@Column(name="PRODUCT_ID")
-	private BigDecimal productId;
+	private BigInteger productId;
 
 	public Picture() {
 	}
 
-	public long getId() {
+	public Long getId() {
 		return this.id;
 	}
 
-	public void setId(long id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
@@ -43,11 +43,11 @@ public class Picture implements Serializable {
 		this.link = link;
 	}
 
-	public BigDecimal getProductId() {
+	public BigInteger getProductId() {
 		return this.productId;
 	}
 
-	public void setProductId(BigDecimal productId) {
+	public void setProductId(BigInteger productId) {
 		this.productId = productId;
 	}
